requirements.txt
python-telegram-bot==20.8
django==5.0
web3==6.15.1
Pillow==10.1.0
python-dotenv==1.0.0
google-cloud-storage==2.13.0